Major change in cmake setup:  * Simox projects are not longer treated as stand-alone-projects, but as subprojects of simox. This avoids lots of confusions that have been ontroduced due to the VirtualRobot stand-alone project.  * cmake build-config and install setup has been improved and is consistent through the whole project.  * switched to version 2.3.0
git-svn-id: https://svn.code.sf.net/p/simox/code/trunk@466 042f3d55-54a8-47e9-b7fb-15903f145c44

+++ GraspPlanning/GraspPlanner/MATPlanner/GridOfMedialSpheres.cpp
@@ -146,7 +146,7 @@ namespace GraspStudio
 		{
 			mainDirection = 'x';
 			//numCells(0) = ceil(ceil(ratio * (numPoints ** (1.0 / 3))) / gridConstant);
-			numCells(0) = (int)ceil(ceil(ratio *pow(numPoints, (1.0f / 3.0f))) / gridConstant);
+			numCells(0) = (int)ceil(ceil(ratio *pow((float)numPoints, (1.0f / 3.0f))) / gridConstant);
 			cellWidth = maxDist / numCells(0);
 			numCells(1) = (int)ceil(diffVect(1) / cellWidth);
 			numCells(2) = (int)ceil(diffVect(2) / cellWidth);
@@ -155,7 +155,7 @@ namespace GraspStudio
 		{
 			mainDirection = 'y';
 			//numCells(1) = ceil(ceil(ratio * (num_points ** (1.0 / 3))) / k_factor);
-			numCells(1) = (int)ceil(ceil(ratio *pow(numPoints, (1.0f / 3.0f))) / gridConstant);
+			numCells(1) = (int)ceil(ceil(ratio *pow((float)numPoints, (1.0f / 3.0f))) / gridConstant);
 			cellWidth = maxDist / numCells(1);
 			numCells(0) = (int)ceil(diffVect(0) / cellWidth);
 			numCells(2) = (int)ceil(diffVect(2) / cellWidth);
@@ -165,7 +165,7 @@ namespace GraspStudio
 		{
 			mainDirection = 'z';
 			//numCells(2) = ceil(ceil(ratio * (numPoints ** (1.0 / 3))) / gridConstant);
-			numCells(2) = (int)ceil(ceil(ratio *pow(numPoints, (1.0f / 3.0f))) / gridConstant);
+			numCells(2) = (int)ceil(ceil(ratio *pow((float)numPoints, (1.0f / 3.0f))) / gridConstant);
 			cellWidth = maxDist / numCells(2);
 			numCells(0) = (int)ceil(diffVect(0) / cellWidth);
 			numCells(1) = (int)ceil(diffVect(1) / cellWidth);
